работаю над картами 2

+++ mods/lg2-0.1.0/src/main/java/com/lostglade/server/MonitorScreenRuntimeModel.java
@@ -89,8 +89,8 @@ record YandexMapsVisualSnapshot(
 		BufferedImage frame,
 		String statusText,
 		String dimensionLabel,
-		int centerX,
-		int centerZ,
+		double centerX,
+		double centerZ,
 		double zoomBlocks,
 		boolean healthy
 ) {
